requirements.txt
anyio==3.5.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==3.0.1
decorator==4.0.11
fastapi==0.75.1
ffmpeg-python==0.2.0
future==0.18.2
idna==3.4
imageio==2.21.2
imageio-ffmpeg==0.4.7
lxml==4.9.1
moviepy==1.0.3
numpy==1.23.2
opencv-python==4.7.0.68
Pillow==9.2.0
proglog==0.1.10
progress==1.6
progressbar==2.5
pydantic==1.9.0
pyjsparser==2.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.17.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
PyPDF2==3.0.1
psutil==5.9.4